+++ clickhouse-http-client/src/main/java/com/clickhouse/client/http/ApacheHttpConnectionImpl.java
@@ -57,6 +57,7 @@ import java.net.ConnectException;
 import java.net.HttpURLConnection;
 import java.net.InetSocketAddress;
 import java.net.Socket;
+import java.net.StandardSocketOptions;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
@@ -97,7 +98,7 @@ public class ApacheHttpConnectionImpl extends ClickHouseHttpConnection {
 
         HttpClientBuilder builder = HttpClientBuilder.create().setConnectionManager(connManager)
                 .disableContentCompression();
-        if (!c.isUseNoProxy() && c.getProxyType() == ClickHouseProxyType.HTTP) {
+        if (c.getProxyType() == ClickHouseProxyType.HTTP) {
             builder.setProxy(new HttpHost(c.getProxyHost(), c.getProxyPort()));
         }
         return builder.build();
@@ -377,8 +378,24 @@ public class ApacheHttpConnectionImpl extends ClickHouseHttpConnection {
             if (config.hasOption(ClickHouseClientOption.SOCKET_TCP_NODELAY)) {
                 builder.setTcpNoDelay(config.getBoolOption(ClickHouseClientOption.SOCKET_TCP_NODELAY));
             }
-            if (!config.isUseNoProxy() && config.getProxyType() == ClickHouseProxyType.SOCKS) {
+            if (config.getProxyType() == ClickHouseProxyType.SOCKS) {
                 builder.setSocksProxyAddress(new InetSocketAddress(config.getProxyHost(), config.getProxyPort()));
+            }
+            if (config.hasOption(ClickHouseClientOption.SOCKET_RCVBUF)) {
+                int bufferSize = config.getIntOption(ClickHouseClientOption.SOCKET_RCVBUF);
+                builder.setRcvBufSize(bufferSize > 0 ? bufferSize : config.getReadBufferSize());
+            } else {
+                int bufferSize = config.getBufferSize();
+                int maxQueuedBuffers = config.getMaxQueuedBuffers();
+                builder.setRcvBufSize(bufferSize * maxQueuedBuffers);
+            }
+            if (config.hasOption(ClickHouseClientOption.SOCKET_SNDBUF)) {
+                int bufferSize = config.getIntOption(ClickHouseClientOption.SOCKET_SNDBUF);
+                builder.setSndBufSize(bufferSize > 0 ? bufferSize : config.getWriteBufferSize());
+            } else {
+                int bufferSize = config.getBufferSize();
+                int maxQueuedBuffers = config.getMaxQueuedBuffers();
+                builder.setSndBufSize(bufferSize * maxQueuedBuffers);
             }
             setDefaultSocketConfig(builder.build());
         }